To properly align the header
I took a different approach to the solution which required editing
the html.
The solution crates separate divs for left and right. I achieved
the same result by using the margin-left:auto style on the button
which created the required separation between left and right items.
The only deviation seems to be in the margin between Logo and links
which I will sort out later.